+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.1.0 [unreleased]
 * Add support for fault sets
+* Add support for background device scanner
 * Bugfix: Do not fail if bootstrap_mdm_name is undef (ie. on a SDC)
 
 ## 3.0.2